Update clean PO files rake tasks
Use new API for gettext v3.

Upstream removes [1] the helper method we were using. So now need to run
the MsgMerge tool directly.

We output to a temp file so we can to correctly set the project ID
version as the previous helper did.

This version of the tool has some difference to the previous version:
1. a bug where new lines where outputted as "\\n" has now been fixed [2]
2. the no-wrap option doesn't prevent new lines from wrapping [3]

This means all our PO files will see updates next time translations are
imported.

[1] https://github.com/ruby-gettext/gettext/commit/f7a9f36#diff-f0acde42603d2c46897ef9cae667e108L70
[2] https://github.com/ruby-gettext/gettext/blob/3.3.5/doc/text/news.md#fixes-16
[3] https://github.com/ruby-gettext/gettext/blob/3.3.5/test/test_po_entry.rb#L709